takes everything related to contacts into a separate module and adds the ability to delete contacts from the address book
